Maintain the correct order of key value pairs
When there are keys that contain dots or spaces and the array version is
requested, they are temporarily replaced and the replacements are then
reverted when converted to an array. Assure that the array maintains the
correct order of key value pairs, when replacements are reverted.
Also introduce changelog file.