Sanitize by converting input values back and forth
To assure that there can't be differences in the array and string
versions returned by the `Query` class, no matter if the instance was
created from string or array, the library now first converts incoming
values back and forth. So, when an instance is created from string, it
first converts it to an array and then again back to a string and vice
versa when an instance is created from an array.